fix(neuron.connection.http): fix error case so return value matches typespec
The definition of the `JSONParseError` struct says that the `response` element is a
`Neuron.Response.t()`, but the code was actually returning a tuple (`{:error,
Neuron.Response.t()}`). With this change the type of `response` matches the typespec.

BREAKING CHANGE: If your code expects the `response` field of `JSONParseError` to be a tuple, it
will break.

fix #59